+++ src/app/products/models/operation-system.enum.ts
@@ -1,0 +1,7 @@
+export enum OperationSystem {
+  Symbian = "Symbian",
+  Android = "Android",
+  IOS = "IOS",
+  Mobian = "Mobian",
+  Bada = "Bada"
+}