+++ src/app/products/models/product.interface.ts
@@ -1,0 +1,11 @@
+import { OperationSystem } from './operation-system.enum';
+
+export interface IProduct {
+  name: string;
+  description: string;
+  price: number;
+  isAvailable: boolean;
+  operationSystem: OperationSystem;
+  availableCount: number;
+  itemsInCart: number;
+}