+++ src/app/shared/shared.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HighlightDirective } from './directives/highlight/highlight.directive';
+
+@NgModule({
+  declarations: [
+    HighlightDirective
+  ],
+  exports: [
+    HighlightDirective
+  ],
+  imports: [
+    CommonModule
+  ]
+})
+export class SharedModule { }